# modules/canvas_headless/openapi.yml
openapi: 3.1.0
info:
  title: Drupal Canvas Headless
  description: API Spec for the Canvas Headless submodule
  version: 0.x

# All routes except the public content endpoint require Drupal session
# authentication. The session cookie name is site-specific, so represent the
# complete Cookie header rather than declaring a fixed cookie name.
security:
  - DrupalSessionCookie: []

# @see https://spec.openapis.org/oas/v3.1.0.html#paths-object
# @see https://spec.openapis.org/oas/v3.1.0.html#path-item-object
paths:
...
  /canvas/content-api/entity:
    get:
      summary: Render one Canvas-managed content entity
      description: |
        Accepts one content entity target and returns its rendered Canvas
        content when Canvas manages the requested entity and view mode.
      parameters:
        - name: type
          in: query
          required: true
          description: Content entity type ID.
          schema:
            type: string
            examples:
              - node
        - name: id
          in: query
          required: true
          description: Content entity ID.
          schema:
            oneOf:
              - type: string
              - type: integer
        - name: viewMode
          in: query
          required: false
          description: Optional content-template view mode.
          schema:
            type: string
            pattern: '^[a-z0-9_]+$'
      responses:
        '200':
          description: Entity render result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanvasEntityResult'
              examples:
                entity:
                  value:
                    content:
                      element: js-example-card
                      props:
                        heading: Example heading
                    managedByCanvas: true
                    entity:
                      entityType: node
                      bundle: article
                      id: '12'
                      uuid: 773942c6-3660-4c50-9a8d-e25966a69bff
                      langcode: en
        '400':
          $ref: '#/components/responses/ProblemDetailsResponse'
        '401':
          $ref: '#/components/responses/ProblemDetailsResponse'
        '403':
          $ref: '#/components/responses/ProblemDetailsResponse'
        '404':
          $ref: '#/components/responses/ProblemDetailsResponse'
        '500':
          $ref: '#/components/responses/ProblemDetailsResponse'
      security: []
...
components:
  schemas:
...
    CanvasEntityResult:
      type: object
      additionalProperties: false
      required: [content, managedByCanvas, entity]
      properties:
        content:
          oneOf:
            - type: object
              additionalProperties: true
            - type: 'null'
        managedByCanvas:
          type: boolean
        entity:
          $ref: '#/components/schemas/RoutedEntityRef'
...
    RoutedEntityRef:
      type: object
      additionalProperties: false
      required: [entityType, bundle, id, uuid, langcode]
      properties:
        entityType:
          type: string
        bundle:
          type: string
        id:
          type: string
        uuid:
          type: string
          format: uuid
        langcode:
          type: string
...
    ProblemDetails:
      type: object
      # RFC 9457 section 3.2 permits extension members, so additional
      # properties must not be forbidden here.
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
  responses:
...
    ProblemDetailsResponse:
      description: RFC 9457 Problem Details error response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    # The remaining routes are handled by Drupal's generic exception handling,
    # which negotiates the response format from the request. Unlike the rest of
    # the Canvas API, these route names are not prefixed with `canvas.api.`, so
    # ApiExceptionSubscriber does not shape them into JSON:API-style errors
    # either.
    # @see \Drupal\canvas\EventSubscriber\ApiExceptionSubscriber::onException()
    # @see \Drupal\Core\EventSubscriber\ExceptionJsonSubscriber::on4xx()
...
  securitySchemes:
    DrupalSessionCookie:
      type: apiKey
      in: header
      name: Cookie
      description: |
        Drupal session cookie header. The cookie name is site-specific and the
        browser normally sends it automatically for same-origin requests.